refactor: Move sized widget code to modules
