+++ tui-popup/src/popup.rs
@@ -1,9 +1,9 @@
 use std::fmt::Debug;
 
-use crate::PopupState;
+use crate::{sized_widget::SizedWidgetRef, PopupState};
 use derive_setters::Setters;
 use ratatui::{
-    prelude::{Buffer, Line, Rect, Style, Text},
+    prelude::{Buffer, Line, Rect, Style},
     symbols::border::Set,
     widgets::{Block, Borders, Clear, StatefulWidgetRef, Widget, WidgetRef},
 };
@@ -48,17 +48,6 @@ pub struct Popup<'content, W: SizedWidgetRef> {
     pub border_style: Style,
 }
 
-/// A trait for widgets that have a fixed size.
-///
-/// This trait allows the popup to automatically size itself based on the size of the body widget.
-/// Implementing this trait for a widget allows it to be used as the body of a popup. You can also
-/// wrap existing widgets in a newtype and implement this trait for the newtype to use them as the
-/// body of a popup.
-pub trait SizedWidgetRef: WidgetRef + Debug {
-    fn width(&self) -> usize;
-    fn height(&self) -> usize;
-}
-
 impl<'content, W: SizedWidgetRef> Popup<'content, W> {
     /// Create a new popup with the given title and body with all the borders.
     ///
@@ -83,49 +72,6 @@ impl<'content, W: SizedWidgetRef> Popup<'content, W> {
             title: Line::default(),
             style: Style::default(),
         }
-    }
-}
-
-impl SizedWidgetRef for &Text<'_> {
-    fn width(&self) -> usize {
-        Text::width(self)
-    }
-
-    fn height(&self) -> usize {
-        Text::height(self)
-    }
-}
-
-impl SizedWidgetRef for &str {
-    fn width(&self) -> usize {
-        Text::from(*self).width()
-    }
-
-    fn height(&self) -> usize {
-        Text::from(*self).height()
-    }
-}
-
-#[derive(Debug)]
-pub struct SizedWrapper<W: Debug> {
-    pub inner: W,
-    pub width: usize,
-    pub height: usize,
-}
-
-impl<W: WidgetRef + Debug> WidgetRef for SizedWrapper<W> {
-    fn render_ref(&self, area: Rect, buf: &mut Buffer) {
-        self.inner.render_ref(area, buf);
-    }
-}
-
-impl<W: WidgetRef + Debug> SizedWidgetRef for SizedWrapper<W> {
-    fn width(&self) -> usize {
-        self.width
-    }
-
-    fn height(&self) -> usize {
-        self.height
     }
 }
